+++ release-notes.md
@@ -213,3 +213,4 @@ Released v0.3.4
 * NEW: The length of the drill in minutes & seconds is shown in the button bar above the timeline.  It also shows the length in counts.
 * NEW: The timeline now shows time of each count in minutes & seconds (mm:ss) on the top.  It continues to show the count on the bottom.
 * NEW: Add Music option on timeline right-click menu.
+* NEW: Keyboard shortcut - Press spacebar to play/stop drill (with music).